+++ stroom-authentication/stroom-authentication-api/src/main/java/stroom/authentication/api/JsonWebKeyFactory.java
@@ -1,0 +1,12 @@
+package stroom.authentication.api;
+
+import org.jose4j.jwk.PublicJsonWebKey;
+
+public interface JsonWebKeyFactory {
+
+    PublicJsonWebKey createPublicKey();
+
+    String asJson(final PublicJsonWebKey publicJsonWebKey);
+
+    PublicJsonWebKey fromJson(final String json);
+}