+++ stroom-authentication/stroom-authentication-impl-db/src/main/java/stroom/authentication/impl/db/JwkDaoImpl.java
@@ -1,16 +1,10 @@
 package stroom.authentication.impl.db;
 
-import org.jose4j.jwk.JsonWebKey;
 import org.jose4j.jwk.PublicJsonWebKey;
-import org.jose4j.jwk.RsaJsonWebKey;
-import org.jose4j.jwk.RsaJwkGenerator;
-import org.jose4j.lang.JoseException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import stroom.authentication.impl.db.jooq.tables.TokenType;
-import stroom.authentication.impl.db.jooq.tables.records.AccountRecord;
-import stroom.authentication.token.JwkDao;
+
+import stroom.authentication.api.JsonWebKeyFactory;
 import stroom.authentication.impl.db.jooq.tables.records.JsonWebKeyRecord;
+import stroom.authentication.token.JwkDao;
 import stroom.authentication.token.Token;
 import stroom.authentication.token.TokenTypeDao;
 import stroom.db.util.JooqUtil;
@@ -24,7 +18,6 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import static stroom.authentication.impl.db.jooq.tables.Account.ACCOUNT;
 import static stroom.authentication.impl.db.jooq.tables.JsonWebKey.JSON_WEB_KEY;
 
 @Singleton
@@ -34,13 +27,16 @@ class JwkDaoImpl implements JwkDao {
     private static final int MAX_KEY_AGE_MS = 1000 * 60 * 60 * 24 * 2;
 
     private final AuthDbConnProvider authDbConnProvider;
+    private final JsonWebKeyFactory jsonWebKeyFactory;
     private final TokenTypeDao tokenTypeDao;
 
     @Inject
     JwkDaoImpl(final AuthDbConnProvider authDbConnProvider,
-               final TokenTypeDao tokenTypeDao) {
+               final TokenTypeDao tokenTypeDao,
+               final JsonWebKeyFactory jsonWebKeyFactory) {
         this.authDbConnProvider = authDbConnProvider;
         this.tokenTypeDao = tokenTypeDao;
+        this.jsonWebKeyFactory = jsonWebKeyFactory;
     }
 
 //    /**
@@ -94,16 +90,11 @@ class JwkDaoImpl implements JwkDao {
 
         // Fetch back all records.
         final List<JsonWebKeyRecord> list = JooqUtil.contextResult(authDbConnProvider, context ->
-                context.selectFrom(JSON_WEB_KEY).fetch());
+                context
+                        .selectFrom(JSON_WEB_KEY)
+                        .fetch());
         return list.stream()
-                .map(r -> {
-                    try {
-                        return RsaJsonWebKey.Factory.newPublicJwk(r.getJson());
-                    } catch (JoseException e) {
-                        LOGGER.error("Unable to create JWK!", e);
-                        throw new RuntimeException(e);
-                    }
-                })
+                .map(jsonWebKeyRecord -> jsonWebKeyFactory.fromJson(jsonWebKeyRecord.getJson()))
                 .collect(Collectors.toList());
     }
 
@@ -114,7 +105,6 @@ class JwkDaoImpl implements JwkDao {
         if (list.size() < 1) {
             addRecord();
         }
-
 
 //        final long oldest = System.currentTimeMillis() - MIN_KEY_AGE_MS;
 //
@@ -137,14 +127,14 @@ class JwkDaoImpl implements JwkDao {
         final long now = System.currentTimeMillis();
         final String uuid = UUID.randomUUID().toString();
         // We need to set up the jwkId so we know which JWTs were signed by which JWKs.
-        final RsaJsonWebKey jwk = createWebKey(uuid);
+        final PublicJsonWebKey publicJsonWebKey = jsonWebKeyFactory.createPublicKey();
         final int tokenTypeId = tokenTypeDao.getTokenTypeId(Token.TokenType.API.getText().toLowerCase());
 
         JooqUtil.context(authDbConnProvider, context -> {
             LOGGER.debug(LambdaLogUtil.message("Creating a {}", JSON_WEB_KEY.getName()));
             final JsonWebKeyRecord record = context.newRecord(JSON_WEB_KEY);
             record.setKeyId(uuid);
-            record.setJson(jwk.toJson(JsonWebKey.OutputControlLevel.INCLUDE_PRIVATE));
+            record.setJson(jsonWebKeyFactory.asJson(publicJsonWebKey));
             record.setCreateTimeMs(now);
             record.setCreateUser("admin");
             record.setUpdateTimeMs(now);
@@ -153,43 +143,6 @@ class JwkDaoImpl implements JwkDao {
             record.setEnabled(true);
             record.store();
         });
-
-
-//        JooqUtil.context(authDbConnProvider, context -> {
-//            try {
-//                // We need to set up the jwkId so we know which JWTs were signed by which JWKs.
-//                String jwkId = UUID.randomUUID().toString();
-//                RsaJsonWebKey jwk = RsaJwkGenerator.generateJwk(2048);
-//                jwk.setKeyId(jwkId);
-//                jwk.setUse("sig");
-//                jwk.setAlgorithm("RS256");
-//
-//                // Persist the public key
-//                JsonWebKeyRecord jwkRecord = new JsonWebKeyRecord();
-//                jwkRecord.setKeyId(jwkId);
-//                jwkRecord.setJson(jwk.toJson(JsonWebKey.OutputControlLevel.INCLUDE_PRIVATE));
-//                jwkRecord.setCreateTimeMs(System.currentTimeMillis());
-//
-//                context.executeInsert(jwkRecord);
-//            } catch (JoseException e) {
-//                LOGGER.error("Unable to create JWK!", e);
-//                throw new RuntimeException(e);
-//            }
-//        });
-    }
-
-    private RsaJsonWebKey createWebKey(final String uuid) {
-        try {
-        // We need to set up the jwkId so we know which JWTs were signed by which JWKs.
-        final RsaJsonWebKey jwk = RsaJwkGenerator.generateJwk(2048);
-        jwk.setKeyId(uuid);
-        jwk.setUse("sig");
-        jwk.setAlgorithm("RS256");
-        return jwk;
-        } catch (final JoseException e) {
-            LOGGER.error("Unable to create JWK!", e);
-            throw new RuntimeException(e);
-        }
     }
 
     private void deleteOldJwkRecords() {