+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v7.0-beta.28] - 2020-04-29
+
+* Issue **#1489** : Reprocess streams feature failing.
+
+* Issue **#1465** : Add default Open ID credentials to allow proxy to be able to authenticate out of the box.
+
 * Issue **#1455** : Fix interactive search.
 
 * Issue **#1471** : Pipeline name not shown on processors/filters in UI.
@@ -13,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Issue **#1491** : Download stream feature failing. 
 
 * Issue **#1433** : StandardKafkaProducer failed when writing XML kafka payloads. 
+
 
 ## [v7.0-beta.27] - 2020-04-27
 
@@ -2074,7 +2082,8 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.27...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.28...HEAD
+[v7.0-beta.28]: https://github.com/gchq/stroom/compare/v7.0-beta.27...v7.0-beta.28
 [v7.0-beta.27]: https://github.com/gchq/stroom/compare/v7.0-beta.26...v7.0-beta.27
 [v7.0-beta.26]: https://github.com/gchq/stroom/compare/v7.0-beta.25...v7.0-beta.26
 [v7.0-beta.25]: https://github.com/gchq/stroom/compare/v7.0-beta.24...v7.0-beta.25